Update SupplyRequest.fsh
Why should the Identifier element be mandatory? I can imagine if requesters would rely on the fhir resources and would not feel it necessary to communicate either keep track of the requests in a proprietary manner, or feels it's not necessary to register a  'private' business identifier in the resource.

What is the rationale behind the proposal to make this element mandatory?